+++ hybrid-array/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Pending
 
 - Added `serde` feature - implements `Serialize` and `Deserialize` for `Array`
+- Added `ArrayExt::try_from_fn`
 
 ## 0.1.0 (2022-05-07)
 - Initial release